Verify feasibility of markdown content in front-matter
It feels a little weird to have to do rendering inside of a rendered template, but I guess it's not so bad?

Note I had to do something a little hacky with the .gitignore for the `src` merged content directory. I'll have to revisit it if I have content files that are not `.md`, but it also might never be an issue.