Add more content from the Ghost site and initial draft support
I had to get rid of the guide-list.liquid snippet because I couldn't figure out how to pass it both a) my collections, and b) a flag indicating which collection (draft or published) to use. This is kind of the same problem I was having before with getting collections into the data cascade. Here, I couldn't get them into the front matter to pass to renderFile, but I decided to keep moving on other stuff.